Corrections + missing part done.
I changed a lot of small things and translated the whole missing part at the bottom.
I'm not sure about everything but I did my best.
Note : Some translation are quite larger than the english version. I'm not sure they could fit in the space dedicated to them.
It has to be tested.